Tests/LibWeb/Text/expected/shadow-root-style-sheets.txt

shadow.styleSheets.length=1
document.styleSheets.length=0
